requirements.txt
fastapi[all]
uvicorn[standard]
simple_salesforce
python-dotenv
firebase
google-cloud-aiplatform